gabriel fix the leaks please

+++ src/utils/utils_parser.c
@@ -6,7 +6,7 @@
 /*   By: dteruya <dteruya@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/05/03 15:03:00 by gcesar-n          #+#    #+#             */
-/*   Updated: 2025/05/23 12:41:23 by dteruya          ###   ########.fr       */
+/*   Updated: 2025/05/23 16:23:45 by dteruya          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,7 +17,7 @@ bool	verify_space(char *str)
 	int	i;
 
 	i = 0;
-	while (i < ft_strlen(str))
+	while (i < (int)ft_strlen(str))
 	{
 		if ((str[i] >= 9 && str[i] <= 13) || str[i] == 32)
 			i++;
